The changes made in this version include:
Removing the global variable page and passing it as an argument to functions.
Adding a new function get_additional_info to extract more details from the output screen.
Modifying the run function to return both name and additional_info.
Printing the name and additional_info in the main function.
Using f-strings to format the output.